day ten
fun one! i spent a while tinkering with part 2 before i realized that i
could "expand" the grid (including the loop) to reveal gaps, and then do
a graph search (DFS in this case) and fill in all of the space that the
loop doesn't enclose.

i looked some stuff up later to discover that there are, of course, some
very elegant solutions that can be mapped onto this problem (shoelace
algorithm and Pick's theorem seemed to be a crowd favorite). i left a
"TODO" for myself to implement a (in my opinion) particularly elegant
and simple ray-firing algorithm to solve part 2 again.

in any case, i'm happy with how my funky "expand and fill" turned out